+++ interfaces/matlab/examples/osqpmatlab.m
@@ -29,6 +29,7 @@ function [x, y, cost, status, iter] = osqpmatlab(problem, warm_start, settings) 
 %     + max_iter: maximum number of iterations
 %     + verbose: verbosity of the solver
 %       TODO: Add these!
+%     + scaled_termination: evaluate termination criteria
 %     + early_terminate: evaluate termination criteria
 %     + early_terminate_interval: interval for evaluating termination criteria
 %