+++ interfaces/matlab/osqp_mex.cpp
@@ -40,14 +40,15 @@ const char* OSQP_SETTINGS_FIELDS[] =
                                 "max_iter",                     //c_int
                                 "eps_abs",                      //c_float
                                 "eps_rel",                      //c_float
-                                "eps_prim_inf",                      //c_float
-                                "eps_dual_inf",                      //c_float
+                                "eps_prim_inf",                 //c_float
+                                "eps_dual_inf",                 //c_float
                                 "alpha",                        //c_float
                                 "delta",                        //c_float
                                 "polish",                       //c_int
                                 "pol_refine_iter",              //c_int
                                 "auto_rho",                     //c_int
                                 "verbose",                      //c_int
+                                "scaled_termination",           //c_int
                                 "early_terminate",              //c_int
                                 "early_terminate_interval",     //c_int
                                 "warm_start"};                  //c_int
@@ -712,6 +713,7 @@ mxArray* copySettingsToMxStruct(OSQPSettings* settings){
   mxSetField(mxPtr, 0, "pol_refine_iter", mxCreateDoubleScalar(settings->pol_refine_iter));
   mxSetField(mxPtr, 0, "auto_rho",        mxCreateDoubleScalar(settings->auto_rho));
   mxSetField(mxPtr, 0, "verbose",         mxCreateDoubleScalar(settings->verbose));
+  mxSetField(mxPtr, 0, "scaled_termination", mxCreateDoubleScalar(settings->scaled_termination));
   mxSetField(mxPtr, 0, "early_terminate", mxCreateDoubleScalar(settings->early_terminate));
   mxSetField(mxPtr, 0, "early_terminate_interval", mxCreateDoubleScalar(settings->early_terminate_interval));
   mxSetField(mxPtr, 0, "warm_start",      mxCreateDoubleScalar(settings->warm_start));
@@ -934,6 +936,7 @@ void copyMxStructToSettings(const mxArray* mxPtr, OSQPSettings* settings){
   settings->pol_refine_iter = (c_int)mxGetScalar(mxGetField(mxPtr, 0, "pol_refine_iter"));
   settings->auto_rho = (c_int)mxGetScalar(mxGetField(mxPtr, 0, "auto_rho"));
   settings->verbose         = (c_int)mxGetScalar(mxGetField(mxPtr, 0, "verbose"));
+  settings->scaled_termination = (c_int)mxGetScalar(mxGetField(mxPtr, 0, "scaled_termination"));
   settings->early_terminate = (c_int)mxGetScalar(mxGetField(mxPtr, 0, "early_terminate"));
   settings->early_terminate_interval = (c_int)mxGetScalar(mxGetField(mxPtr, 0, "early_terminate_interval"));
   settings->warm_start      = (c_int)mxGetScalar(mxGetField(mxPtr, 0, "warm_start"));
@@ -967,6 +970,8 @@ void copyUpdatedSettingsToWork(const mxArray* mxPtr ,OsqpData* osqpData){
     (c_int)mxGetScalar(mxGetField(mxPtr, 0, "pol_refine_iter")));
   osqp_update_verbose(osqpData->work,
     (c_int)mxGetScalar(mxGetField(mxPtr, 0, "verbose")));
+  osqp_update_scaled_termination(osqpData->work,
+    (c_int)mxGetScalar(mxGetField(mxPtr, 0, "scaled_termination")));
   osqp_update_early_terminate(osqpData->work,
     (c_int)mxGetScalar(mxGetField(mxPtr, 0, "early_terminate")));
   osqp_update_early_terminate_interval(osqpData->work,