+++ src/util.c
@@ -84,21 +84,25 @@ void print_setup_header(const OSQPData *data, const OSQPSettings *settings) {
     c_print("max_iter = %i\n", (int)settings->max_iter);
 
     if (settings->early_terminate)
-        c_print("          early_terminate: active (interval %i)\n", (int)settings->early_terminate_interval);
+        c_print("          early_terminate: on (interval %i)\n", (int)settings->early_terminate_interval);
     else
-        c_print("          early_terminate: inactive\n");
+        c_print("          early_terminate: off \n");
     if (settings->scaling)
-        c_print("          scaling: active\n");
+        c_print("          scaling: on, ");
     else
-        c_print("          scaling: inactive\n");
+        c_print("          scaling: off, ");
+    if (settings->scaled_termination)
+        c_print("scaled_termination: on\n");
+    else
+        c_print("scaled_termination: off\n");
     if (settings->warm_start)
-        c_print("          warm start: active\n");
+        c_print("          warm start: on, ");
     else
-        c_print("          warm start: inactive\n");
+        c_print("          warm start: off, ");
     if (settings->polish)
-        c_print("          polish: active\n");
+        c_print("polish: on\n");
     else
-        c_print("          polish: inactive\n");
+        c_print("polish: off\n");
     c_print("\n");
 }
 
@@ -201,6 +205,7 @@ void set_default_settings(OSQPSettings * settings) {
         settings->verbose = VERBOSE;     /* print output */
         #endif
 
+        settings->scaled_termination = SCALED_TERMINATION;     /* Evaluate scaled termination criteria*/
         settings->early_terminate = EARLY_TERMINATE;     /* Evaluate termination criteria */
         settings->early_terminate_interval = EARLY_TERMINATE_INTERVAL;     /* Evaluate termination at certain interval */
         settings->warm_start = WARM_START;     /* x equality constraint scaling: 1e-3 */
@@ -228,6 +233,7 @@ OSQPSettings * copy_settings(OSQPSettings * settings){
     new->pol_refine_iter = settings->pol_refine_iter;
     new->auto_rho = settings->auto_rho;
     new->verbose = settings->verbose;
+    new->scaled_termination = settings->scaled_termination;
     new->early_terminate = settings->early_terminate;
     new->early_terminate_interval = settings->early_terminate_interval;
     new->warm_start = settings->warm_start;